// client/app/api/accounting/model/UserModel.ts
/**
 * GuanPlus.AccountingFirm.WebApi
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import * as models from './models';

/**
 * 系统用户
 */
export interface UserModel {
    /**
     * 用户id
     */
    id?: string;

    /**
     * 编码
     */
    code?: string;

    /**
     * 名称
     */
    name?: string;

    /**
     * 电子邮件
     */
    email?: string;

    /**
     * 电话号码
     */
    phoneNumber?: string;

    /**
     * 头像地址
     */
    accountPhotoUrl?: string;

    /**
     * 职位
     */
    position?: string;

    /**
     * 地址
     */
    address?: models.AddressModel;

    /**
     * 角色
     */
    roles?: Array<models.RoleModel>;

    /**
     * 用户帐套列表
     */
    accountBooks?: Array<models.AccountBookModel>;

    /**
     * 账套Id
     */
    accountBookId?: string;

    /**
     * 是否管理员
     */
    admin?: boolean;

    /**
     * 是否激活
     */
    active?: boolean;

    /**
     * 是否可用
     */
    isEnable?: string;

}